Add comments to test. Make abstract base for tests in prep for VB
